change double to int

+++ inclass/mem_test.cpp
@@ -1,7 +1,7 @@
 #include "std_lib_facilities.h"
 
 int main(){
-	int* arr = new double[10000];
+	double* arr = new double[10000];
 	
 	for(int i = 0; i < 10000; i++){
 		arr[i] = i;
